fix: corrige erro de tipos no INSERT de itens da NF de importação (patch 0064)
Erro em produção logo após o patch 0064 (visto ao abrir a NF 143003 e
outras): "inconsistent types deduced for parameter \$5" (42P08, detail
"integer versus numeric"). Causa: o parâmetro \$5 (quantidade_esperada,
numeric) também era usado dentro de "CASE WHEN \$6 THEN \$5 ELSE 0 END"
pra decidir quantidade_recebida - o literal "0" no ELSE é inteiro por
padrão, e o Postgres exige um tipo único e consistente pra todo uso do
mesmo parâmetro na query, gerando o conflito.

Corrigido com cast explícito: "ELSE 0::numeric END" - agora as duas
aparições de \$5 concordam em numeric.

Confirmado o erro real via logs de runtime do Vercel (boxer-wms-api),
rastreado até nf-importacao.js:204 no deployment que já estava em
produção (dpl_H6i9rwEWR142N1dGzT567QLjCx1m, commit bba24054).

+++ wms-api/routes/nf-importacao.js
@@ -206,7 +206,7 @@ router.get('/:id/itens', async (req, res) => {
                     (nota_id, item_erp_id, sku, descricao, quantidade_esperada, quantidade_recebida,
                      recebido_automaticamente, unidade, valor_unitario,
                      peso_liquido_kg, peso_bruto_kg, comprimento_cm, largura_cm, altura_cm, volume_m3)
-                 VALUES ($1, $2, $3, $4, $5, CASE WHEN $6 THEN $5 ELSE 0 END, $6, $7, $8, $9, $10, $11, $12, $13, $14)
+                 VALUES ($1, $2, $3, $4, $5, CASE WHEN $6 THEN $5 ELSE 0::numeric END, $6, $7, $8, $9, $10, $11, $12, $13, $14)
                  ON CONFLICT (item_erp_id) DO UPDATE SET quantidade_esperada = EXCLUDED.quantidade_esperada
                  RETURNING id, quantidade_recebida, recebido_automaticamente`,
                 [
